requirements.txt
scipion-pyworkflow
scipion-em
scipion-em-phenix
scipion-em-continuousflex